Feature: I made a property page. If the "Residencies" a attribute in the navbar is clicked, it redirects to the property page,
i also updated the slider.json, to add more dummy properites, as well as i installed "react-query", as i wanted to use it for error handling but didnt end up using it. changed the property.css file a bit.

I also added, in the page.tsx, i made the logo image an a attribute so that users can navigate bac to the homepage